feat(cli): add check-latest-version utility function
This commit adds a new utility function called `checkLatestVersion` to the `utils` directory in the `cli` package. The function is responsible for checking if the user is using the latest stable version of the CLI. It compares the current version of the CLI with the latest version available and displays a message if an update is available.

The `checkLatestVersion` function uses the following helper functions:
- `getCliVersion`: This function retrieves the latest version of the CLI from the npm registry.
- `getManifest`: This function retrieves the manifest file of the CLI, which contains information about the CLI, including its name and current version.

The `checkLatestVersion` function performs the following steps:
1. Retrieves the current version of the CLI from the manifest file.
2. Retrieves the latest version of the CLI from the npm registry using the `getCliVersion` function.
3. Compares the latest version with the current version.
4. If the latest version is different from the current version, a message is displayed using the `note` function from the `@clack/prompts` package. The message informs the user that a new version is available and provides instructions on how to update the CLI.

This utility function is useful for notifying users about new versions of the CLI and encouraging them to update to the latest version to benefit from new features and bug fixes.